Fixed issue [ 1915246 ] 64bit compile: feps returns exactly zero. slamch_ was declared incorrectly as "double", should be "float".
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@3156 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/LAPACK.hpp
@@ -26,7 +26,7 @@
 extern "C" {
   double dlamch_(char * CMACH);
   
-  double slamch_(char * CMACH);
+  float slamch_(char * CMACH);
 
   void dgesvx_(char* FACT, char* TRANS, int * N, int * NRHS, 
 	       double *A, int * LDA, double * AF, int * LDAF,
